// GradeSession.Client/src/shared/api/generated/model/inviteResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */
import type { SessionParticipantRole } from './sessionParticipantRole';

export interface InviteResponse {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  committeeSlotId?: string | null;
  role?: SessionParticipantRole;
  createdAt?: string;
  /** @nullable */
  claimedByUserId?: string | null;
  /** @nullable */
  claimedAt?: string | null;
  /** @nullable */
  revokedAt?: string | null;
}