// GradeSession.Client/src/shared/api/generated/model/issuedInviteResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */
import type { SessionParticipantRole } from './sessionParticipantRole';

export interface IssuedInviteResponse {
  /** @nullable */
  inviteId?: string | null;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  committeeSlotId?: string | null;
  role?: SessionParticipantRole;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  shortName?: string | null;
  /** @nullable */
  token?: string | null;
}